+++ packages/app/hivecommand-native/src/views/setup/index.tsx
@@ -17,6 +17,7 @@ export const SetupView = (props: any) => {
 
     const [activeIndex, setActiveIndex] = useState(0);
 
+    console.log({globalState})
 
     const steps = [
         {
@@ -58,10 +59,17 @@ export const SetupView = (props: any) => {
                 console.log({networkLayout: globalState?.networkLayout})
                 axios.post(`http://localhost:${8484}/setup`, {
                     config: {
-                        host: globalState?.networkLayout?.iotEndpoint,
-                        user: globalState?.networkLayout?.iotUser,
-                        pass: globalState?.networkLayout?.iotToken,
-                        exchange: globalState?.networkLayout?.iotSubject
+                        tags: globalState?.controlLayout?.tags,
+                        types: globalState?.controlLayout?.types,
+                        opcuaServer: authState?.opcuaServer,
+                        iot: {
+                            host: globalState?.networkLayout?.iotEndpoint,
+                            user: globalState?.networkLayout?.iotUser,
+                            pass: globalState?.networkLayout?.iotToken,
+                            exchange: globalState?.networkLayout?.iotSubject
+                        },
+                        subscriptionMap: globalState?.subscriptionMap,
+                        deviceMap: globalState?.deviceMap
                     }
                 }).then((data) => {
                     console.log("OPCUA PROVISIONED");