+++ packages/communication/opc-ethernet-ip-bridge/src/bridge/index.test.ts
@@ -1,0 +1,18 @@
+describe('ENIP Bridge', () => {
+
+    it('Can load Array type with arraySize and format as native type', () => {
+
+    })
+
+    it('Can write Array type values', () => {
+        
+    })
+
+    it('Can load bit_string values', () => {
+
+    })
+
+    it('Can write bit_string values', () => {
+
+    })
+})