+++ packages/communication/opc-mqtt/src/index.ts
@@ -33,25 +33,36 @@ export class MQTTPublisher {
     }
 
     //Subscribe to changes requested by other entities
-    async subscribe(onMessage: (message: ConsumeMessage | null) => void){
+    async subscribe(onMessage: (message: {routingKey?: string, userId?: string, messageContent: any}) => Promise<void>){
         const generatedQueue = await this.channel?.assertQueue(`device:${this.options.user}`)
         if(!generatedQueue) return;
 
-        // await this.channel?.bindQueue(generatedQueue.queue, this.options.exchange, key);
-
-        this.channel?.consume(generatedQueue.queue, onMessage);
+        this.channel?.consume(generatedQueue.queue, async (message) => {
+            if(!message) return;
+            const routingKey = message?.fields.routingKey;
+            const userId : string | undefined = message?.properties.userId;
+            const messageContent = JSON.parse(message?.content.toString() || '{error: "No message content"}');
+    
+            try{
+                await onMessage({routingKey, userId, messageContent})
+            }catch(e){
+                console.error("Error leading to nack", message)
+                return this.channel?.nack(message)
+            }
+            return this.channel?.ack(message)
+        });
     }
 
     //Publish current state to other entities
     async publish(key: string, dataType: string, value: any){
-        
+        if(!value) return null;
+
         console.log("Publishing ", key, dataType, value)
         if(value.BYTES_PER_ELEMENT != undefined){
-            console.log("Publishing typed array", value);
             value = Array.from(value)
         }
 
-        this.channel?.publish(
+        return this.channel?.publish(
             this.options.exchange, 
             key, 
             Buffer.from(JSON.stringify({