+++ packages/core-ui/command-surface/src/views/control.tsx
@@ -182,14 +182,14 @@ export default () => {
                     if (Number.isNaN(val)) {
                         val = 0;
                     }
-                    return val.toFixed(2);
+                    return val % 1 != 0 ? val.toFixed(2) : val;
                 }) : (() => {
                     let val = parseFloat(value || 0);
 
                     if(Number.isNaN(val)) {
                         val = 0;
                     }
-                    return val.toFixed(2);
+                    return val % 1 != 0 ? val.toFixed(2) : val;
                 })()
             default:
                 console.log({ type })
@@ -315,10 +315,8 @@ export default () => {
 
 	// console.log({ hmiNodes, operatingMode, operatingModes })
 
+	console.log({values, hmi: hmi?.nodes})
 
-	const infoBubble = useMemo(() => {
-		
-	}, [])
 
 	return (
 		<Box sx={{ flex: 1, display: 'flex', flexDirection: "row", position: 'relative' }}>