+++ packages/core/command-local-sidecar/src/utils/index.ts
@@ -1,0 +1,13 @@
+
+export const load_exports = (code: string) => {
+
+    const _require = (name: string) => {
+        console.log("Requires", name)
+    }
+
+    const exports : any = {};
+    const module = { exports };
+    const func = new Function("require", "module", "exports", code);
+    func(_require, module, exports);
+    return module.exports;
+}