+++ packages/clients/local-client/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.172](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.171...v1.4.15-alpha.172) (2023-02-14)
+
+**Note:** Version bump only for package @hive-command/local-client
+
+
+
+
+
+## [1.4.15-alpha.171](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.170...v1.4.15-alpha.171) (2023-02-14)
+
+**Note:** Version bump only for package @hive-command/local-client
+
+
+
+
+
+## [1.4.15-alpha.170](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.169...v1.4.15-alpha.170) (2023-02-13)
+
+**Note:** Version bump only for package @hive-command/local-client
+
+
+
+
+
+## [1.4.15-alpha.169](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.168...v1.4.15-alpha.169) (2023-02-13)
+
+**Note:** Version bump only for package @hive-command/local-client
+
+
+
+
+
+## [1.4.15-alpha.168](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.167...v1.4.15-alpha.168) (2023-02-13)
+
+**Note:** Version bump only for package @hive-command/local-client
+
+
+
+
+
+## [1.4.15-alpha.167](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.166...v1.4.15-alpha.167) (2023-02-12)
+
+**Note:** Version bump only for package @hive-command/local-client
+
+
+
+
+
 ## [1.4.15-alpha.166](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.165...v1.4.15-alpha.166) (2023-02-12)
 
 **Note:** Version bump only for package @hive-command/local-client